Piece bit field can now be overwritten on peers
Added an overwrite function. It expect a binary of the same size as the
existing bit field.

When a peer send a bit field we can now overwrite its pieces set, which
will be initialized as empty. When the fast extension is getting
implemented we can just call the has all function.